+++ src/main-handlers/action/parse-config.js
