+++ src/main-handlers/api/method-partial.md
