+++ src/main-handlers/lambda/index.js
@@ -6,11 +6,11 @@ const modules = {
 
 module.exports = templatePath => {
   console.log('INFO: retrieving lambda functions data from serverless.yml file...')
-  const meta = require('../get-sls.js')(modules, __dirname)
-  const data = require('../get-data.js')(modules, meta, 'lambda')
+  const meta = require('../../helpers/get-sls.js')(modules, __dirname)
+  const data = require('../../helpers/get-data.js')(modules, meta, 'lambda')
   console.log('SUCCESS: extracted lambda functions meta data!')
   console.log('INFO: generating documentation...')
-  const docs = require('../build-docs.js')(modules, data, templatePath, 'lambda')
+  const docs = require('../../helpers/build-docs.js')(modules, data, templatePath, 'lambda')
   console.log('SUCCESS: documentation successfully generated!')
   modules.fs.writeFileSync('README.md', docs, 'utf8')
 }
