+++ src/main-handlers/lambda/process-meta.js
