+++ src/main-handlers/package/get-main.js
